re-organize univariate finite files
