+++ src/MLJBase.jl
@@ -246,10 +246,10 @@ include("info_dict.jl")
 
 include("interface/data_utils.jl")
 include("interface/model_api.jl")
-include("interface/univariate_finite.jl")
 
-include("distributions.jl")
-include("univariate_finite_array.jl")
+include("univariate_finite/types.jl")
+include("univariate_finite/methods.jl")
+include("univariate_finite/arrays.jl")
 
 include("machines.jl")
