+++ src/univariate_finite/arrays.jl
@@ -49,15 +49,3 @@ function Base.Broadcast.broadcasted(
     return broadcast(pdf, u, class)
 end
 
-# convenient broadasting of pdf for multiples classes:
-function Base.Broadcast.broadcasted(
-    ::typeof(pdf),
-    u::AbstractArray{UnivariateFinite{S,V,R,P},N},
-    C::AbstractVector{<:V}) where {S,V,R,P,N}
-
-    ret = Array{P,N+1}(undef, size(u)..., length(C))
-    for i in eachindex(C)
-        ret[fill(:,N)...,i] = broadcast(pdf, u, C[i])
-    end
-    return ret
-end
