+++ test/runtests.jl
@@ -38,15 +38,19 @@ print("\r                                           \r")
 
 @testset "misc" begin
 #    @test include("utilities.jl")
-    @test include("distributions.jl")
 #    @test include("parameter_inspection.jl")
 #    @test include("info_dict.jl")
 #    @test include("static.jl")
 end
 
-# @testset "interface" begin
-#     @test include("interface/interface.jl")
-# end
+@testset "interface" begin
+     @test include("interface/interface.jl")
+end
+
+@testset "univariate finite" begin
+     @test include("univariate_finite/methods.jl")
+     @test include("univariate_finite/arrays.jl")
+end
 
 # @testset "measures" begin
 #     @test include("measures/measures.jl")
