+++ test/univariate_finite/obsolete.jl
@@ -10,78 +10,6 @@ rng = StableRNG(111)
 n   = 10
 c   = 3
 
-@testset "Constr:binary" begin
-    scores  = rand(rng, n)
-    support = ["class1", "class2"]
-
-    u = UV(support, scores)
-    @test length(u) == n
-    @test size(u) == (n,)
-    @test u.scores[1] == scores[1]
-    @test u.support[1] == support[1]
-    @test u.support[2] == support[2]
-
-    # autosupport
-    u = UV(scores)
-    @test u.support[1] == :negative
-    @test u.support[2] == :positive
-
-    # errors
-    scores = [-1,0,1]
-    @test_throws DomainError UV(scores)
-end
-
-@testset "Constr:multi" begin
-    P   = rand(rng, n, c)
-    @test_throws DomainError UV(P)
-
-    P ./= sum(P, dims=2)
-    u   = UV(P)
-
-    @test length(u) == n
-    @test length(u.support) == c
-    @test size(u) == (n,)
-    @test u.support[1] == :class_1
-    @test u.support[2] == :class_2
-    @test u.support[3] == :class_3
-end
-
-@testset "Get and Set" begin
-    # binary
-    s = rand(rng, n)
-    u = UV(s)
-
-    @test u[1] isa UnivariateFinite
-    @test all(MLJBase.classes(u[1]) .== u.support)
-
-    u[1] = 0.3
-    @test pdf(u[1], u.support[2]) == 0.3
-    u[1:2] = [0.3,0.2]
-    @test pdf(u[2], :positive) == 0.2
-
-    @test_throws DomainError (u[1] = -0.2)
-    @test_throws DomainError (u[1:2] = [-0.1,0.1])
-
-    # multiclass
-    P   = rand(rng, n, c)
-    P ./= sum(P, dims=2)
-    u   = UV(P)
-
-    @test u[1] isa UnivariateFinite
-    @test all(MLJBase.classes(u[1]) .== u.support)
-
-    u[1] = [0.3, 0.5, 0.2]
-    @test pdf(u[1], u.support[1]) == 0.3
-    @test pdf(u[1], u.support[2]) == 0.5
-    @test pdf(u[1], u.support[2]) == 0.5
-    u[1:2] = [0.3 0.4 0.3; 0.2 0.2 0.6]
-    @test pdf(u[1], u.support[2]) == 0.4
-    @test pdf(u[2], u.support[3]) == 0.6
-
-    @test_throws DomainError (u[1] = [-0.2,0.3,0.9])
-    @test_throws DomainError (u[1:2] = rand(2,3))
-end
-
 @testset "Mode/Pdf" begin
     # binary
     rng = StableRNG(668)
